Poprawa SQLi do tworzenia tabel i naprawa błędów w create'ach.

+++ pl.atena.aj.be.ebay/src/main/resources/sql/auction.sql
@@ -3,6 +3,6 @@ CREATE TABLE `auction` (
   `name`             varchar(256)         DEFAULT NULL COMMENT 'Nazwa aukcji',
   `description`      varchar(512)         DEFAULT NULL COMMENT 'Opis aukcji',
   `end_date`         date                 DEFAULT NULL COMMENT 'Data koñca aukcji',
-  `final_price`      double(4,2) unsigned DEFAULT NULL COMMENT 'Cena koñcowa',
+  `final_price`      double(10,2) unsigned DEFAULT NULL COMMENT 'Cena koñcowa',
   PRIMARY KEY (`auction_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
